@@ -1,6 +1,6 @@
 module Fast-MBTiler
 
-go 1.15
+go 1.16
 
 require (
 	github.com/antonfisher/nested-logrus-formatter v1.3.0
